Agregando funciones con data de los productos

+++ src/components/Cart.jsx
@@ -1,31 +1,46 @@
 import { React, useContext } from "react";
 import { CartContext, deleteItem } from "../Context/CartContext";
+import { Link } from "react-router-dom";
 
 function Cart() {
-  const { cart, deleteItem } = useContext(CartContext);
+  const { cart, deleteItem, totalPrecioUnidad, deleteCart } = useContext(CartContext);
+
+  if (cart.length === 0)
+    return (
+      <>
+        <h1>
+          Aun no has agregado productos al carrito, te invitamos a <Link to="/">Volver al catalogo</Link> para que selecciones alguno
+        </h1>
+      </>
+    );
 
   return (
-    <div className="cartContainer">
-      {cart.map((itemInCart, index) => {
-        return (
-          <div className="containerInternoCart">
-            <div className="containerImg">
-              <img className="imgProductCart" src={`${itemInCart.img}`} alt="" />
+    <>
+      <div className="cartContainer">
+        {cart.map((itemInCart, index) => {
+          return (
+            <div className="containerInternoCart">
+              <div className="containerImg">
+                <img className="imgProductCart" src={`${itemInCart.img}`} alt="" />
+              </div>
+              <div className="titleAndRemove">
+                <p>{itemInCart.title}</p>
+                <a
+                  onClick={() => {
+                    deleteItem(index);
+                  }}>
+                  Remove Item
+                </a>
+              </div>
+              <p>{itemInCart.cantidad}</p>
+              <h3></h3>
             </div>
-            <div className="titleAndRemove">
-              <p>{itemInCart.title}</p>
-              <a
-                onClick={() => {
-                  deleteItem(index);
-                }}>
-                Remove Item
-              </a>
-            </div>
-            <p>{itemInCart.cantidad}</p>
-          </div>
-        );
-      })}
-    </div>
+          );
+        })}
+      </div>
+      <h3>Total:${totalPrecioUnidad()}</h3>
+      <button onClick={deleteCart}>Vaciar carrito</button>
+    </>
   );
 }
